refactor: move layout related components to AppComponents
