+++ src/components/SharedComponents/AppLayout/AppLayout.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import Layout, { LayoutProps } from "@/components/SharedComponents/Layout";
+import Layout, { LayoutProps } from "@/components/AppComponents/Layout";
 
 export type AppLayoutProps = LayoutProps;
 const AppLayout: FC<AppLayoutProps> = (props) => {
